Pull

+++ maintenance/checksync.sh
@@ -81,6 +81,7 @@ function sync_check() {
     # check if current to within 1.5 minutes
     if ((TIMEDIF <= 90 && TIMEDIF >= -90))
     then echo -e " The blockchain is almost certainly synced.\n"
+        echo -e " $(date +%m.%d.%Y_%H:%M:%S) : ${lightgreen}Masternode ${PROJECT}_n${i} synced completely ${nocolor}" | tee -a "$LOGFILE"
         SYNCED="yes"
     else echo -e " That's the same as${yellow} $((($(date +%s)-NEWEST)/3600)) hours${nocolor} or${yellow} $((($(date +%s)-NEWEST)/86400)) days${nocolor} behind the present.\n"
         SYNCED="no"
@@ -124,6 +125,12 @@ function check_blocksync() {
         if [ "$SYNCED" = "yes" ]; then echo -e "              ${lightgreen}Masternode synced${nocolor}\n" ; break
         else echo -e "${white} Blockchain is ${lightred}not yet synced${nocolor}; will check again in 10 seconds${nocolor}\n"
             echo -e " I have been checking this masternode for:${lightcyan} $SECONDS seconds${nocolor}\n"
+            # if clonesyncing, display warning not to interrupt it
+            if [ -e $INSTALLDIR/temp/clonesyncing ]
+            then echo -e " ${lightred}Clonesync_all in progress; DO NOT INTERRUPT THIS PROCESS!!${nocolor}"
+                echo -e " ${lightred}Bootstrap will resume once your first blockchain is synced.${nocolor}\n"
+            else :
+            fi
             # insert a little humor
             curl -s "http://api.icndb.com/jokes/random" | jq '.value.joke'
             echo -e "\n"
@@ -132,12 +139,11 @@ function check_blocksync() {
         fi
     done
 
-    if [ "$SYNCED" = "no" ]; then echo -e "${lightred} Masternode did not sync in the allowed time${nocolor}\n"
+    if [ "$SYNCED" = "no" ]; then echo -e "${lightred} Masternode n$i did not sync in the allowed time${nocolor}\n" | tee -a "$LOGFILE"
         # exit the script because syncing did not occur
         rm -rf $INSTALLDIR/getinfo_n${i} --force
         exit
-
-else : ; fi
+    else : ; fi
 
     #   create file to signal that this blockchain is synced (I moved this to the cronchecksync)
     #   echo -e " Setting flag at: $INSTALLDIR/temp/${PROJECT}_n${i}_synced\n"
